Errores con los scripts arreglados
