+++ Assets/Scripts/DamageIndicator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9a6e15a4d1b53ad4cadf3c0fa508a5a1
+guid: 51006703cc23a6144bbbf8e37ab7c380
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
