+++ Assets/Scripts/DamageIndicatorDirection.cs
@@ -16,40 +16,41 @@ public class DamageIndicatorDirection : DamageIndicator
         _charHP.onHPChangeBy += OnHarm;
     }
 
-    private void OnHarm(float damage,GameObject source)
+    private void OnHarm(float damage, GameObject source)
     {
         OnHarm(damage);
-        if(_nodePool.Count <= 0)
+        if (_nodePool.Count <= 0)
         {
-            DamageIndicatorDirectionNode instantiatedNode = Instantiate(_baseNode,transform);
+            DamageIndicatorDirectionNode instantiatedNode = Instantiate(_baseNode, transform);
             instantiatedNode.damageSourcePosition = source.transform.position;
             instantiatedNode.onLifeTimeExpire += _nodePool.Push;
             instantiatedNode.victimTransform = _charOrigin;
-            if(instantiatedNode.lifeTime == 0)
+            if (instantiatedNode.lifeTime == 0)
             {
                 instantiatedNode.lifeTime = feedbackDuration;
             }
-            
+
         }
         else
         {
             _nodePool.Pop().gameObject.SetActive(true);
-        if(_nodePool.Count <= 0)
-        {
-            DamageIndicatorDirectionNode instantiatedNode = Instantiate(_baseNode,transform);
-            instantiatedNode.damageSourcePosition = source.transform.position;
-            instantiatedNode.onLifeTimeExpire += _nodePool.Push;
-            instantiatedNode.victimTransform = _charOrigin;
-            if(instantiatedNode.lifeTime == 0)
+            if (_nodePool.Count <= 0)
             {
-                instantiatedNode.lifeTime = feedbackDuration;
+                DamageIndicatorDirectionNode instantiatedNode = Instantiate(_baseNode, transform);
+                instantiatedNode.damageSourcePosition = source.transform.position;
+                instantiatedNode.onLifeTimeExpire += _nodePool.Push;
+                instantiatedNode.victimTransform = _charOrigin;
+                if (instantiatedNode.lifeTime == 0)
+                {
+                    instantiatedNode.lifeTime = feedbackDuration;
+                }
+
             }
-            
+            else
+            {
+                _nodePool.Pop().gameObject.SetActive(true);
+            }
+
         }
-        else
-        {
-            _nodePool.Pop().gameObject.SetActive(true);
-        }
-        
     }
 }
