+++ Assets/Scripts/DamageIndicatorDirection.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 16221d11e92f1fc45925173e81f8251f
+guid: bf747ede37cee954198c2e005150acf2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
