+++ Assets/Scripts/DamageIndicatorDirectionNode.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4217e9bdb795d784282c3aa7b38c02db
+guid: 9fb510cd83f5b464ebfc0afcdba8996e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
