--- Assets/Scripts/UI/DamageIndicator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6b7a478139af4f44892fe788cb90dc22
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
