--- Assets/Scripts/UI/DamageIndicatorDirection.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 868ecc2bafdcc7a4a8817f14c8c1a49f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
